requirements.txt
groq>=0.10.0
python-dotenv>=1.0.1
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
